+++ crates/filter/src/lib.rs
@@ -13,7 +13,7 @@ use anyhow::Result;
 use rayon::prelude::*;
 
 use icon::Icon;
-use matcher::{MatchScope, Matcher};
+use matcher::{Bonus, MatchScope, Matcher};
 
 pub use self::source::Source;
 pub use self::worker::iterator::dyn_run;
@@ -61,6 +61,11 @@ impl FilterContext {
 
     pub fn match_scope(mut self, match_scope: MatchScope) -> Self {
         self.matcher = self.matcher.set_match_scope(match_scope);
+        self
+    }
+
+    pub fn bonuses(mut self, bonuses: Vec<Bonus>) -> Self {
+        self.matcher = self.matcher.set_bonuses(bonuses);
         self
     }
 }