+++ crates/filter/src/worker/iterator.rs
@@ -285,13 +285,13 @@ fn dyn_collect_number(
 /// Returns the ranked results after applying fuzzy filter given the query string and a list of candidates.
 pub fn dyn_run<I: Iterator<Item = SourceItem>>(
     query: &str,
-    source: Source<I>,
     FilterContext {
         icon,
         number,
         winwidth,
         matcher,
     }: FilterContext,
+    source: Source<I>,
 ) -> Result<()> {
     let query: Query = query.into();
     if let Some(number) = number {
@@ -370,6 +370,7 @@ mod tests {
         let mut total_lines_created: usize = 0;
         dyn_run(
             "abc",
+            FilterContext::default().number(Some(100)),
             Source::List(
                 std::iter::repeat_with(|| {
                     bytes = bytes.reverse_bits().rotate_right(3).wrapping_add(1);
@@ -398,7 +399,6 @@ mod tests {
                 })
                 .take(usize::max_value() >> 8),
             ),
-            FilterContext::default().number(Some(100)),
         )
         .unwrap()
     }