+++ crates/maple_cli/src/command/filter.rs
@@ -119,8 +119,8 @@ impl Filter {
         } else {
             filter::dyn_run::<std::iter::Empty<_>>(
                 &self.query,
-                self.generate_source(),
                 FilterContext::new(icon, number, winwidth, matcher),
+                self.generate_source(),
             )?;
         }
         Ok(())