+++ crates/maple_cli/src/command/grep/forerunner.rs
@@ -1,5 +1,16 @@
-use super::*;
-use crate::process::light::CommandEnv;
+use std::path::PathBuf;
+use std::process::Command;
+
+use anyhow::Result;
+use clap::Parser;
+
+use utility::is_git_repo;
+
+use super::{RG_ARGS, RG_EXEC_CMD};
+use crate::app::Params;
+use crate::process::light::{CommandEnv, LightCommand};
+use crate::process::BaseCommand;
+use crate::utils::{send_response_from_cache, SendResponse};
 
 #[derive(Parser, Debug, Clone)]
 pub struct RipGrepForerunner {
@@ -54,15 +65,14 @@ impl RipGrepForerunner {
             return Ok(());
         }
 
-        let mut std_cmd = StdCommand::new(RG_ARGS[0]);
+        // Can not use StdCommand as it joins the args which does work somehow.
+        let mut cmd = Command::new(RG_ARGS[0]);
         // Do not use --vimgrep here.
-        std_cmd.args(&RG_ARGS[1..]);
+        cmd.args(&RG_ARGS[1..]);
 
         if let Some(ref dir) = self.cmd_dir {
-            std_cmd.current_dir(dir);
+            cmd.current_dir(dir);
         }
-
-        let mut cmd = std_cmd.into_inner();
 
         let mut light_cmd = LightCommand::new(
             &mut cmd,