+++ crates/maple_cli/src/process/rstd.rs
@@ -10,7 +10,7 @@ pub fn collect_stdout(cmd: &mut Command) -> std::io::Result<Vec<u8>> {
     if !cmd_output.status.success() && !cmd_output.stderr.is_empty() {
         return Err(std::io::Error::new(
             std::io::ErrorKind::Other,
-            format!("an error occured: {:?}", cmd_output.stderr),
+            String::from_utf8_lossy(&cmd_output.stderr),
         ));
     }
 
@@ -74,6 +74,7 @@ impl StdCommand {
         collect_stdout(&mut self.0)
     }
 
+    #[allow(unused)]
     pub fn args<I, S>(&mut self, args: I) -> &mut Self
     where
         I: IntoIterator<Item = S>,