+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Fix the incompatiblity issue between vim-signature and vim-clap popup window. #817
 - Escape the file name for filer sink. #822
 - Seperate the vista impl for tags provider completely. #827
+- Fix the regression of command `ripgrep-forerunner`.
 
 ## Improved